12 Part 2 ok for real data

+++ AdventOfCode/AdventOfCode/2023/Day12/Day12_Part2.cs
@@ -4,7 +4,7 @@
     {
         private readonly Day12 _day12 = new();
 
-        public int Run(string input)
+        public long Run(string input)
         {
             return _day12.Run(input, true);
         }
